refactorizacion primera forma de crear compound components
