+++ 8-compound-component-pattern/src/component-patterns/components/ProductTitle.tsx
@@ -1,0 +1,14 @@
+import { useContext } from "react"
+import { ProductContext } from "./ProductCard"
+
+import styles from "../styles/styles.module.css";
+
+export const ProductTitle = ({ title}:{title?:string}) => { // se obliga el titulo, si hay mas propiedades uso una interface
+  const {product} = useContext(ProductContext)
+
+  return (
+    <span className={styles.productDescription}>
+      { title ? title : product.title}
+    </span>
+  )
+}
