+++ 8-compound-component-pattern/src/component-patterns/interfaces/interfaces.ts
@@ -1,0 +1,21 @@
+import { ReactElement } from "react";
+
+export interface ProductCardProps {
+  product: Product;
+  children?: ReactElement | ReactElement []
+}
+
+export interface Product  {
+  id: string,
+  title: string,
+  img?: string
+}
+
+// context api:
+
+export interface ProductContextProps {
+  counter: number; 
+  increaseBy: (value: number) => void,
+  product: Product
+
+}
